+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### New features
+
+* [#83](https://github.com/tboox/xmake/issues/83): Add `add_csnippet` and `add_cxxsnippet` into `option` for detecting some compiler features.
+
 ## v2.1.4
 
 ### New features
@@ -271,6 +275,10 @@
 # 更新日志
 
 ## master (开发中)
+
+### 新特性
+
+* [#83](https://github.com/tboox/xmake/issues/83): 添加 `add_csnippet`，`add_cxxsnippet`到`option`来检测一些编译器特性
 
 ## v2.1.4